chore: create separate sharing component with workspace sharing hook
