+++ site/src/modules/groups.ts
@@ -1,15 +1,21 @@
-import type { Group, ReducedUser, User } from "api/typesGenerated";
+import type {
+	Group,
+	ReducedUser,
+	User,
+	WorkspaceUser,
+} from "api/typesGenerated";
 
-type UserOrGroupAutocompleteValue = User | ReducedUser | Group | null;
+/**
+ * Union of all user-like types that can be distinguished from Group.
+ */
+type UserLike = User | ReducedUser | WorkspaceUser;
 
 /**
  * Type guard to check if the value is a Group.
  * Groups have a "members" property that users don't have.
  */
-export const isGroup = (
-	value: UserOrGroupAutocompleteValue,
-): value is Group => {
-	return value !== null && typeof value === "object" && "members" in value;
+export const isGroup = (value: UserLike | Group): value is Group => {
+	return "members" in value;
 };
 
 /**
