+++ site/src/pages/WorkspaceSettingsPage/WorkspaceSharingPage/WorkspaceSharingPage.tsx
@@ -1,16 +1,10 @@
 import { checkAuthorization } from "api/queries/authCheck";
-import {
-	setWorkspaceGroupRole,
-	setWorkspaceUserRole,
-	workspaceACL,
-} from "api/queries/workspaces";
-import { ErrorAlert } from "components/Alert/ErrorAlert";
-import { displaySuccess } from "components/GlobalSnackbar/utils";
 import { Link } from "components/Link/Link";
 import type { WorkspacePermissions } from "modules/workspaces/permissions";
 import { workspaceChecks } from "modules/workspaces/permissions";
+import { useWorkspaceSharing } from "modules/workspaces/WorkspaceSharingForm/useWorkspaceSharing";
 import type { FC } from "react";
-import { useMutation, useQuery, useQueryClient } from "react-query";
+import { useQuery } from "react-query";
 import { docs } from "utils/docs";
 import { pageTitle } from "utils/page";
 import { useWorkspaceSettings } from "../WorkspaceSettingsLayout";
@@ -18,32 +12,17 @@ import { WorkspaceSharingPageView } from "./WorkspaceSharingPageView";
 
 const WorkspaceSharingPage: FC = () => {
 	const workspace = useWorkspaceSettings();
-	const queryClient = useQueryClient();
+	const sharing = useWorkspaceSharing(workspace);
 
-	const workspaceACLQuery = useQuery(workspaceACL(workspace.id));
 	const checks = workspaceChecks(workspace);
 	const permissionsQuery = useQuery<WorkspacePermissions>({
 		...checkAuthorization({ checks }),
 	});
 	const permissions = permissionsQuery.data;
-
-	const addUserMutation = useMutation(setWorkspaceUserRole(queryClient));
-	const updateUserMutation = useMutation(setWorkspaceUserRole(queryClient));
-	const removeUserMutation = useMutation(setWorkspaceUserRole(queryClient));
-
-	const addGroupMutation = useMutation(setWorkspaceGroupRole(queryClient));
-	const updateGroupMutation = useMutation(setWorkspaceGroupRole(queryClient));
-	const removeGroupMutation = useMutation(setWorkspaceGroupRole(queryClient));
-
 	const canUpdatePermissions = Boolean(permissions?.updateWorkspace);
 
-	const mutationError =
-		addUserMutation.error ??
-		updateUserMutation.error ??
-		removeUserMutation.error ??
-		addGroupMutation.error ??
-		updateGroupMutation.error ??
-		removeGroupMutation.error;
+	const error =
+		sharing.error ?? permissionsQuery.error ?? sharing.mutationError;
 
 	return (
 		<div className="flex flex-col gap-12 max-w-screen-md">
@@ -60,80 +39,21 @@ const WorkspaceSharingPage: FC = () => {
 				</div>
 			</header>
 
-			{workspaceACLQuery.isError && (
-				<ErrorAlert error={workspaceACLQuery.error} />
-			)}
-			{permissionsQuery.isError && (
-				<ErrorAlert error={permissionsQuery.error} />
-			)}
-			{Boolean(mutationError) && <ErrorAlert error={mutationError} />}
-
 			<WorkspaceSharingPageView
 				workspace={workspace}
-				workspaceACL={workspaceACLQuery.data}
+				workspaceACL={sharing.workspaceACL}
 				canUpdatePermissions={canUpdatePermissions}
-				onAddUser={async (user, role, reset) => {
-					await addUserMutation.mutateAsync({
-						workspaceId: workspace.id,
-						userId: user.id,
-						role,
-					});
-					displaySuccess("User added to workspace successfully!");
-					reset();
-				}}
-				isAddingUser={addUserMutation.isPending}
-				onUpdateUser={async (user, role) => {
-					await updateUserMutation.mutateAsync({
-						workspaceId: workspace.id,
-						userId: user.id,
-						role,
-					});
-					displaySuccess("User role updated successfully!");
-				}}
-				updatingUserId={
-					updateUserMutation.isPending
-						? updateUserMutation.variables?.userId
-						: undefined
-				}
-				onRemoveUser={async (user) => {
-					await removeUserMutation.mutateAsync({
-						workspaceId: workspace.id,
-						userId: user.id,
-						role: "",
-					});
-					displaySuccess("User removed successfully!");
-				}}
-				onAddGroup={async (group, role, reset) => {
-					await addGroupMutation.mutateAsync({
-						workspaceId: workspace.id,
-						groupId: group.id,
-						role,
-					});
-					displaySuccess("Group added to workspace successfully!");
-					reset();
-				}}
-				isAddingGroup={addGroupMutation.isPending}
-				onUpdateGroup={async (group, role) => {
-					await updateGroupMutation.mutateAsync({
-						workspaceId: workspace.id,
-						groupId: group.id,
-						role,
-					});
-					displaySuccess("Group role updated successfully!");
-				}}
-				updatingGroupId={
-					updateGroupMutation.isPending
-						? updateGroupMutation.variables?.groupId
-						: undefined
-				}
-				onRemoveGroup={async (group) => {
-					await removeGroupMutation.mutateAsync({
-						workspaceId: workspace.id,
-						groupId: group.id,
-						role: "",
-					});
-					displaySuccess("Group removed successfully!");
-				}}
+				error={error}
+				onAddUser={sharing.addUser}
+				isAddingUser={sharing.isAddingUser}
+				onUpdateUser={sharing.updateUser}
+				updatingUserId={sharing.updatingUserId}
+				onRemoveUser={sharing.removeUser}
+				onAddGroup={sharing.addGroup}
+				isAddingGroup={sharing.isAddingGroup}
+				onUpdateGroup={sharing.updateGroup}
+				updatingGroupId={sharing.updatingGroupId}
+				onRemoveGroup={sharing.removeGroup}
 			/>
 		</div>
 	);
